feat: renamed files and menu rendering catches missing menu ids
